+++ src/ext-test/java/org/opentripplanner/ext/fares/impl/FareModelForTest.java
@@ -13,6 +13,8 @@ import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.network.Route;
 import org.opentripplanner.transit.model.site.FareZone;
 import org.opentripplanner.transit.model.site.RegularStop;
+import org.opentripplanner.transit.service.StopModel;
+import org.opentripplanner.transit.service.StopModelBuilder;
 
 public class FareModelForTest {
 
@@ -23,33 +25,35 @@ public class FareModelForTest {
     .of(FeedScopedId.ofNullable("F2", "other-feed-zone"))
     .build();
 
-  static RegularStop AIRPORT_STOP = RegularStop
-    .of(id("airport"))
+  private static final StopModelBuilder STOP_MODEL_BUILDER = StopModel.of();
+
+  static RegularStop AIRPORT_STOP = STOP_MODEL_BUILDER
+    .regularStop(id("airport"))
     .withCoordinate(new WgsCoordinate(1, 1))
     .addFareZones(AIRPORT_ZONE)
     .withName(new NonLocalizedString("Airport"))
     .build();
 
-  static RegularStop CITY_CENTER_A_STOP = RegularStop
-    .of(id("city-center-a"))
+  static RegularStop CITY_CENTER_A_STOP = STOP_MODEL_BUILDER
+    .regularStop(id("city-center-a"))
     .withCoordinate(new WgsCoordinate(1, 2))
     .addFareZones(CITY_CENTER_ZONE)
     .withName(new NonLocalizedString("City center: stop A"))
     .build();
-  static RegularStop CITY_CENTER_B_STOP = RegularStop
-    .of(id("city-center-b"))
+  static RegularStop CITY_CENTER_B_STOP = STOP_MODEL_BUILDER
+    .regularStop(id("city-center-b"))
     .withCoordinate(new WgsCoordinate(1, 3))
     .addFareZones(CITY_CENTER_ZONE)
     .withName(new NonLocalizedString("City center: stop B"))
     .build();
-  static RegularStop SUBURB_STOP = RegularStop
-    .of(id("suburb"))
+  static RegularStop SUBURB_STOP = STOP_MODEL_BUILDER
+    .regularStop(id("suburb"))
     .withCoordinate(new WgsCoordinate(1, 4))
     .withName(new NonLocalizedString("Suburb"))
     .build();
 
-  static RegularStop OTHER_FEED_STOP = RegularStop
-    .of(FeedScopedId.ofNullable("F2", "other-feed-stop"))
+  static RegularStop OTHER_FEED_STOP = STOP_MODEL_BUILDER
+    .regularStop(FeedScopedId.ofNullable("F2", "other-feed-stop"))
     .withCoordinate(new WgsCoordinate(1, 5))
     .withName(new NonLocalizedString("Other feed stop"))
     .addFareZones(OTHER_FEED_ZONE)