+++ src/main/java/org/opentripplanner/transit/model/site/RegularStop.java
@@ -5,6 +5,7 @@ import java.time.ZoneId;
 import java.util.Collection;
 import java.util.Objects;
 import java.util.Set;
+import java.util.function.IntSupplier;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.locationtech.jts.geom.Point;
@@ -39,7 +40,7 @@ public final class RegularStop
 
   RegularStop(RegularStopBuilder builder) {
     super(builder);
-    this.index = INDEX_COUNTER.getAndIncrement();
+    this.index = builder.createIndex();
     this.platformCode = builder.platformCode();
     this.url = builder.url();
     this.timeZone = builder.timeZone();
@@ -52,8 +53,8 @@ public final class RegularStop
     }
   }
 
-  public static RegularStopBuilder of(FeedScopedId id) {
-    return new RegularStopBuilder(id);
+  public static RegularStopBuilder of(FeedScopedId id, IntSupplier indexCounter) {
+    return new RegularStopBuilder(id, indexCounter);
   }
 
   @Override