+++ src/test/java/org/opentripplanner/routing/algorithm/GraphRoutingTest.java
@@ -80,9 +80,8 @@ public abstract class GraphRoutingTest {
 
     protected Builder() {
       var deduplicator = new Deduplicator();
-      var stopModel = new StopModel();
       graph = new Graph(deduplicator);
-      transitModel = new TransitModel(stopModel, deduplicator);
+      transitModel = new TransitModel(new StopModel(), deduplicator);
       vertexFactory = new VertexFactory(graph);
       vehicleParkingHelper = new VehicleParkingHelper(graph);
     }
@@ -222,13 +221,11 @@ public abstract class GraphRoutingTest {
         .build();
     }
 
-    public RegularStop stopEntity(
-      String id,
-      double latitude,
-      double longitude,
-      boolean noTransfers
-    ) {
-      var stopBuilder = TransitModelForTest.stop(id).withCoordinate(latitude, longitude);
+    RegularStop stopEntity(String id, double latitude, double longitude, boolean noTransfers) {
+      var stopModelBuilder = transitModel.getStopModel().withContext();
+      var testModel = new TransitModelForTest(stopModelBuilder);
+
+      var stopBuilder = testModel.stop(id).withCoordinate(latitude, longitude);
       if (noTransfers) {
         stopBuilder.withParentStation(
           Station
@@ -241,7 +238,7 @@ public abstract class GraphRoutingTest {
       }
 
       var stop = stopBuilder.build();
-      transitModel.mergeStopModels(StopModel.of().withRegularStop(stop).build());
+      transitModel.mergeStopModels(stopModelBuilder.withRegularStop(stop).build());
       return stop;
     }