+++ docs/Changelog.md
@@ -34,6 +34,11 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Fix sort order bug in optimized transfers [#5446](https://github.com/opentripplanner/OpenTripPlanner/pull/5446)
 - Siri file loader [#5460](https://github.com/opentripplanner/OpenTripPlanner/pull/5460)
 - Calculate CO₂ emissions of itineraries [#5278](https://github.com/opentripplanner/OpenTripPlanner/pull/5278)
+- Update NeTEx Java Model 2.0.15 [#5466](https://github.com/opentripplanner/OpenTripPlanner/pull/5466)
+- Migrate to Java 21 [#5421](https://github.com/opentripplanner/OpenTripPlanner/pull/5421)
+- Add Roadmap setup docs [#5468](https://github.com/opentripplanner/OpenTripPlanner/pull/5468)
+- Interpolate increasing stop times for GTFS-RT cancelled trips [#5348](https://github.com/opentripplanner/OpenTripPlanner/pull/5348)
+- Remove itineraries outside the search window in arriveBy search [#5433](https://github.com/opentripplanner/OpenTripPlanner/pull/5433)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.4.0 (2023-09-13)